[Fix][UI Next][V1.0.0-Alpha] Fix the datasource display error in SQL.
