+++ dolphinscheduler-ui-next/src/views/projects/task/components/node/fields/index.ts
@@ -30,7 +30,6 @@ export { useTaskType } from './use-task-type'
 export { useProcessName } from './use-process-name'
 export { useChildNode } from './use-child-node'
 export { useTargetTaskName } from './use-target-task-name'
-export { useDatasourceType } from './use-datasource-type'
 export { useDatasource } from './use-datasource'
 export { useSqlType } from './use-sql-type'
 export { useProcedure } from './use-procedure'
